GH-113: Refactor navigation and overlay components for simplification.
this close #113
Replaced `NavigationBar` implementation with a streamlined `AppBar` component and removed `DrawerToggleButton`. Renamed `OverlayButtons` to `BottomOverlay` for better clarity and updated all relevant imports and usages. This improves code readability and component modularity.
