+++ systems/web/src/pages/core-values.astro
@@ -4,10 +4,10 @@ import HeadingSection from "../components/core-values/react/HeadingSection.tsx"
 import ValueSection, {prepareCoreValuesProps} from "../components/core-values/react/ValueSection.tsx"
 
 import cms from "../cms.ts";
-import NavigationBar, {DrawerToggleButton} from '../components/react/NavigationBar.tsx';
+import NavigationBar from '../components/react/NavigationBar.tsx';
 import Layout from '../layouts/Layout.astro';
 import {Page, Main} from "../components/react/PageContainer.tsx"
-import OverlayButtons from "../components/react/OverlayButtons.tsx"
+import BottomOverlay from "../components/react/BottomOverlay.tsx"
 const {page} = await cms.gqlRequest(Query).then(data => data.data)
 const title = page.title
 const values =  await prepareCoreValuesProps(page.values.values);
@@ -17,13 +17,12 @@ const values =  await prepareCoreValuesProps(page.values.values);
         description={"Explore my core values—Evolution, Interactive Development, Communication, and Flexibility—and how they shape my approach as a professional engineer. Learn how these principles guide my work, collaboration, and continuous growth in the tech industry."}
 >
     <Page>
-        <NavigationBar title={title}>
-            <DrawerToggleButton client:load slot="drawer" title={title}/>
-        </NavigationBar>
+        <NavigationBar client:load title={title}/>
+
         <Main>
             <HeadingSection whatIsCoreValues={page.whatAreCoreValues}/>
             {values.map((value,index) => <ValueSection value={value} variants={index %2 === 0 ? 'right' : 'left'} />)}
-            <OverlayButtons client:load/>
+            <BottomOverlay client:load/>
         </Main>
     </Page>
 </Layout>
