+++ systems/web/src/pages/faq.astro
@@ -1,10 +1,10 @@
 ---
 
-import NavigationBar, {DrawerToggleButton} from '../components/react/NavigationBar.tsx';
+import NavigationBar from '../components/react/NavigationBar.tsx';
 import Layout from '../layouts/Layout.astro';
 import {Page, Main} from "../components/react/PageContainer"
 import cms from "../cms.ts";
-import OverlayButtons from "../components/react/OverlayButtons";
+import BottomOverlay from "../components/react/BottomOverlay";
 import TableOfContentSection from "../components/faq/react/TableOfContentSection.tsx"
 import QuestionAnswerListingSection from "../components/faq/react/QuestionAnswerListingSection.tsx"
 
@@ -17,9 +17,8 @@ const questions = page.questions
 ---
 <Layout title={title} description="Explore the portfolio of a software engineer with a passion for iterative development, team collaboration, and continuous improvement. Learn about their experience, skills, and approach to tackling challenges in software development.">
     <Page>
-        <NavigationBar title={title}>
-            <DrawerToggleButton client:load slot="drawer" title={title}/>
-        </NavigationBar>
+        <NavigationBar client:load title={title}/>
+
         <Main sx={{
                 display: 'flex',
                 flexDirection: 'column',
@@ -27,7 +26,7 @@ const questions = page.questions
                 }}>
             <TableOfContentSection questions={questions}/>
             <QuestionAnswerListingSection questions={questions}/>
-            <OverlayButtons client:load/>
+            <BottomOverlay client:load/>
         </Main>
     </Page>
 </Layout>
