+++ systems/web/src/pages/skills.astro
@@ -1,11 +1,11 @@
 ---
 
-import NavigationBar, {DrawerToggleButton} from '../components/react/NavigationBar.tsx';
+import NavigationBar from '../components/react/NavigationBar.tsx';
 import Layout from '../layouts/Layout.astro';
 import {Page, Main} from "../components/react/PageContainer"
 import SkillSection from "../components/skills/react/SkillSection.tsx"
 import cms from "../cms.ts";
-import OverlayButtons from "../components/react/OverlayButtons";
+import BottomOverlay from "../components/react/BottomOverlay";
 
 import Query from "../components/skills/query.graphql"
 const {page} = await cms.gqlRequest(Query).then(data => data.data)
@@ -15,16 +15,15 @@ const skills = page.skillsRef.sections;
 ---
 <Layout title={title} description={"I am an experienced software engineer specializing in backend and frontend development, automated testing, cloud infrastructure, and Agile collaboration. Skilled in Node.js, React, CI/CD, cloud services (AWS, GCP), and testing frameworks like Jest and Cypress. Strong expertise in building scalable applications and working in Agile teams."}>
     <Page>
-        <NavigationBar title={title}>
-            <DrawerToggleButton client:load slot="drawer" title={title}/>
-        </NavigationBar>
+        <NavigationBar client:load title={title}/>
+
         <Main sx={{
             display: 'flex',
             flexDirection: 'column',
             gap: 1
         }}>
             {skills.map(skill => (<SkillSection section={skill}/>))}
-            <OverlayButtons client:load/>
+            <BottomOverlay client:load/>
         </Main>
     </Page>
 </Layout>
